Modify migration to exclude html-tidy on annotated instances
